requirements.txt
bokeh==3.4.0
brotli==1.1.0
bytecode==0.15.1
cached-property==1.5.2
diffusers
einops==0.8.0
ftfy==6.2.0
graphviz==0.20.3
h2==4.1.0
importlib-resources==6.1.3
ipykernel==6.29.3
ipyplot==1.1.2
ipython-genutils==0.2.0
levenshtein==0.25.1
opencv-python==4.9.0.80
ortools==9.9.3963
p-tqdm==1.4.0
peft==0.7.0
pickleshare==0.7.5
pip-chill==1.0.3
pkgutil-resolve-name==1.3.10
plotly==5.22.0
pynvml==11.5.0
pysocks==1.7.1
pyspice==1.5
scikit-image==0.22.0
scikit-learn==1.4.2
seaborn==0.13.2
shapely==2.0.4
tabulate==0.9.0
tensorboard==2.16.2
tensorboardx==2.6.2.2
torchaudio==2.3.0
torchinfo==1.8.0
torchmetrics==1.3.2
torchvision==0.18.0
types-tabulate==0.9.0.20240106
typing-utils==0.1.0
torch-tb-profiler
wandb
nbformat
ncps
optuna
optuna-dashboard
mambapy